Add hack to handle smaller border blocks
Border bocks can be smaller than regular blocks, which currently leads
to unexpected EOF errors. A quick and dirty way to handle this for the
common case is to just ignore the error. This is obviously not ideal,
but other errors should be rather rare, esp. since HTTP and loading
errors are handled separately.