Fix effective block size computation on bounday blocks
Depending on the dataset size, the boundary blocks might not be of full
size. This was taken into account already, but the computation didn't
respect the voxel offset. This is fixed now and the minimum part of the
boundary is now used to shift the block start accordingly.

Due to the voxel offset, the minimum of the voxel bounds can be
negative. This is why now an i64 data type is used for the computation.